split core components from cirrus.cli (mostly)
