+++ src/cirrus/core/collection_meta.py
