+++ src/cirrus/core/components/base/step_function.py
@@ -2,7 +2,7 @@ import logging
 
 from .. import files
 from .component import Component
-from cirrus.cli.utils.yaml import NamedYamlable
+from cirrus.core.utils.yaml import NamedYamlable
 
 
 logger = logging.getLogger(__name__)
