+++ src/cirrus/core/exceptions.py
