+++ src/cirrus/core/resources/outputs.py
