+++ src/cirrus/core/utils/console.py
