+++ src/cirrus/core/utils/misc.py
